requirements.txt
# === Core Backend ===
fastapi==0.104.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
python-multipart==0.0.6
python-dotenv==1.0.0

# === Authentication ===
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# === Speech-to-Text (OpenAI Cloud) ===
openai==1.58.1

# === Database ===
psycopg2-binary==2.9.11
